% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{anova.mpmm}
\alias{anova.mpmm}
\title{perform likelihood ratio tests on 2 or more mpmm model objects}
\usage{
\method{anova}{mpmm}(m, ...)
}
\arguments{
\item{m}{an model object with class mpmm}

\item{...}{additional mpmm model objects}
}
\description{
perform likelihood ratio tests on 2 or more mpmm model objects
}
