% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmm-package.R
\docType{data}
\name{ellie.ice.short}
\alias{ellie.ice.short}
\title{foieGras-filtered Southern elephant seal Argos satellite data with
environmental covariates (4 individuals)}
\format{
.RData
}
\description{
Example elephant seal Argos tracking data with environmental
covariates. Data were sourced from the Integrated Marine Observing System
(IMOS) - IMOS is supported by the Australian Government through the National
Collaborative Research Infrastructure Strategy and the Super Science
Initiative.
}
\keyword{data}
