% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{fixef}
\alias{fixef}
\alias{fixef.mpmm}
\title{Extract fixed-effects estimates}
\usage{
\method{fixef}{mpmm}(object, ...)
}
\arguments{
\item{object}{a fitted mpmm object}

\item{\dots}{optional additional arguments. not implemented}
}
\value{
an object of class \code{fixef.mpmm} comprising a list of component (\code{cond})
}
\description{
Extract Fixed Effects
}
\details{
Extract fixed effects from a fitted \code{mpmm} model.

The print method for \code{fixef.mpmm} object \emph{only displays non-trivial components}
}
\examples{

}
\keyword{models}
