% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_control.R
\name{inner_control}
\alias{inner_control}
\title{Control Values for \code{mpmm}}
\usage{
inner_control(...)
}
\arguments{
\item{...}{control parameters for the inner (newton) optimizer}
}
\value{
Returns a list with components:
 \item{\code{control}}{list of control parameters for inner optimizer}
}
\description{
\code{inner_control} allows specification of control parameters for the
inner (newton) optimization used by \code{mpmm}
}
\details{
See \code{\link{TMB::MakeADFun}} and \code{\link{TMB::newton}}
for details and available options. Adapted from S. Wotherspoon
\url{https://github.com/SWotherspoon/RWalc/blob/master/R/RWalc.R}
}
\examples{
fit <- mpmm(~ ice + (ice | id),
data = ellie.ice.short,
inner_control = list(
    maxit = 1500,
    tol = 1e-04))
}
\seealso{
\code{\link{TMB::MakeADFun}}, \code{\link{TMB::newton}}
}
