% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{logLik.mpmm}
\alias{logLik.mpmm}
\title{Extract log-likelihood}
\usage{
\method{logLik}{mpmm}(object, ...)
}
\arguments{
\item{object}{an mpmm model fit object}

\item{...}{additional arguments to be ignored}
}
\description{
extract log-likelihood from an mpmm fit object
}
