% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm.r
\name{mpm}
\alias{mpm}
\title{Random Walk with autocorrelation Filter}
\usage{
mpm(d, formula = NA, ts = 12, nu = 5, span = 0.4, optim = c("nlminb",
  "optim"), verbose = FALSE)
}
\arguments{
\item{d}{a data frame of observations (see details)}

\item{formula}{a right-hand-side regression formula}

\item{ts}{time-step in hours (used only when fitting to Argos data)}

\item{nu}{degrees of freedom for t-distributed measurement error (used only when fitting to Argos data)}

\item{span}{degree of loess smoothing for location state starting values (used only when fitting to Argos data)}

\item{optim}{numerical optimizer}

\item{verbose}{report progress during minimization}
}
\value{
a list with components
\item{\code{states}}{a dataframe of estimated states}
\item{\code{fitted}}{a dataframe of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{input dataframe}
\item{\code{tmb}}{the tmb object}
\item{\code{opt}}{the object returned by the optimizer}
}
\description{
Move Persistence Model
}
\details{
Fit a random walk with time-varying move persistence to location data
with or without measurement error

The input track is given as a dataframe where each row is an
observed location and columns
\describe{
\item{'date'}{observation time (POSIXct,GMT),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude,}
\item{'lc'}{ARGOS location class.}
}
}
