% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmm-package.R
\name{mpmm-package}
\alias{mpmm-package}
\title{\pkg{mpmm}}
\description{
fit movement persistence mixed-effects models to animal location data
}
\references{
Jonsen ID, McMahon CR, Patterson TA, et al. (2019) Movement responses to environment: fast inference of variation among southern elephant seals with a mixed effects model. Ecology. 100(1):e02566 https://doi.org/10.1002/ecy.2566
}
\seealso{
mpmm
}
\author{
Ian Jonsen
}
\keyword{mpmm}
