% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmm_control.R
\name{mpmm_control}
\alias{mpmm_control}
\title{Control Values for \code{mpmm}.}
\usage{
mpmm_control(
  optim = c("nlminb", "optim"),
  method = c("BFGS", "L-BFGS-B"),
  lower = NULL,
  upper = NULL,
  REML = FALSE,
  profile = FALSE,
  verbose = 2,
  ...
)
}
\arguments{
\item{optim}{the numerical optimizer used in the fit}

\item{method}{optimization method to be used - one of "BFGS" or "L-BFGS-B"
for bounded optimization (default \code{lower} and \code{upper} bounds are
used if not specified}

\item{lower}{a list of named parameter lower bounds, if NULL then built in
defaults are used when \code{method = "L-BFGS-B"}, otherwise ignored}

\item{upper}{a list of named parameter upper bounds, if NULL then built in
defaults are used when \code{method = "L-BFGS-B"}, otherwise ignored}

\item{REML}{logical; whether to use REML (TRUE) or maximum likelihood}

\item{profile}{logical; option to improve speed and convergence by using
REML parameter estimates as initial values for ML optimization}

\item{verbose}{integer; report progress during minimization: 0 = silent;
1 = optimizer trace; 2 = parameter trace (default))}

\item{...}{control parameters for the chosen optimizer}
}
\value{
Returns a list with components
  \item{\code{optim}}{the name of the numerical optimizer as a
  string, "nlminb" or "optim"}
  \item{\code{method}}{optimization method to be used}
  \item{\code{lower}}{named list of lower parameter bounds}
  \item{\code{upper}}{named list of upper parameter bounds}
  \item{\code{REML}}{whether REML is to be used in place of ML}
  \item{\code{profile}}{whether to enhance convergence robustness}
  \item{\code{verbose}}{level of tracing information to be reported}
  \item{\code{control}}{list of control parameters for the optimizer}
}
\description{
\code{mpmm_control} selects the numerical minimizer, method, associated
control parameters, parameter bounds, and likelihood estimation (REML or ML)
used by \code{mpmm}.
}
\details{
The optimizer used to minimize the objective function is
selected by the \code{optim} argument.  Additional control
parameters specific to the chosen optimizer are specified via the
dots argument.  See \code{\link{nlminb}} and \code{\link{optim}}
for available options. Adapted from S. Wotherspoon
\url{https://github.com/SWotherspoon/RWalc/blob/master/R/RWalc.R}
}
\seealso{
\code{\link{nlminb}}, \code{\link{optim}}.
}
