% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mpmm.r
\name{plot.mpmm}
\alias{plot.mpmm}
\title{plot}
\usage{
\method{plot}{mpmm}(m, label = FALSE)
}
\arguments{
\item{m}{a fitted object of class mpmm}

\item{label}{add id labels to random effects}
}
\description{
Visualise fixed and random relationships
}
