% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_re.r
\name{plot_re}
\alias{plot_re}
\title{plot}
\usage{
plot_re(m)
}
\arguments{
\item{x}{a fitted object of class mpmm}

\item{page}{1 = plot all terms on a single page, 0 otherwise}
}
\description{
Plot fixed and random components
}
