% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.mpmm.r
\name{residuals.mpmm}
\alias{residuals.mpmm}
\title{residuals}
\usage{
\method{residuals}{mpmm}(m, method = "oneStepGaussianOffMode",
  trace = FALSE, parallel = TRUE, ncores = detectCores() - 1)
}
\arguments{
\item{m}{a fitted object of class mpmm}

\item{method}{Character naming the method to calculate one-step-ahead residuals}
}
\value{
a list with components
\item{\code{res}}{a tibble with one-step-ahead residuals for longitude and latitude}
}
\description{
Extract one-step-ahead residuals
}
\details{
Wrapper function for TMB::oneStepPredict function that calculates one-step-ahead residuals, which are residuals that account for temporal correlation in latent states.
}
\examples{
\dontrun{
data(ellie.ice)
fit <- mpmm(~ ice + (1 | id), data = ellie.ice)
summary(fit)
residuals(fit)
}
}
\references{
Thygesen, U. H., C. M. Albertsen, C. W. Berg, K. Kristensen, and A. Neilsen. 2017. Validation of ecological state space models using the Laplace approximation. Environmental and Ecological Statistics 24:317–339.
}
