% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mpmm.r
\name{summary.mpmm}
\alias{summary.mpmm}
\title{object summaries}
\usage{
\method{summary}{mpmm}(object, ...)
}
\arguments{
\item{object}{an mpmm fit object}

\item{...}{additional arguments to be ignored}
}
\description{
return a summary of an mpmm fit object
}
