% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBoneStepPredict.R
\name{TMBoneStepPredict}
\alias{TMBoneStepPredict}
\title{Calculate one-step-ahead (OSA) residuals for a latent variable model.}
\usage{
TMBoneStepPredict(
  obj,
  observation.name = NULL,
  data.term.indicator = NULL,
  method = c("oneStepGaussianOffMode", "fullGaussian", "oneStepGeneric",
    "oneStepGaussian", "cdf"),
  subset = NULL,
  conditional = NULL,
  discrete = NULL,
  discreteSupport = NULL,
  range = c(-Inf, Inf),
  seed = 123,
  parallel = FALSE,
  ncores = 2,
  trace = TRUE,
  reverse = (method == "oneStepGaussianOffMode"),
  ...
)
}
\arguments{
\item{obj}{Output from \code{MakeADFun}.}

\item{observation.name}{Character naming the observation in the template.}

\item{data.term.indicator}{Character naming an indicator data variable in the template (not required by all methods - see details).}

\item{method}{Method to calculate OSA (see details).}

\item{subset}{Index vector of observations that will be added one by one during OSA. By default \code{1:length(observations)} (with \code{conditional} subtracted).}

\item{conditional}{Index vector of observations that are fixed during OSA. By default the empty set.}

\item{discrete}{Are observations discrete? (assumed FALSE by default)}

\item{discreteSupport}{Possible outcomes of discrete distribution (\code{method="oneStepGeneric"} only).}

\item{range}{Possible range of the observations.}

\item{seed}{Randomization seed (discrete case only). If \code{NULL} the RNG seed is untouched by this routine.}

\item{parallel}{Run in parallel using the \code{parallel} package?}

\item{ncores}{number of threads to run in parallel. Ignored if parallel = FALSE}

\item{trace}{Trace progress?}

\item{reverse}{Do calculations in opposite order to improve stability ? (currently enabled by default for \code{oneStepGaussianOffMode} method only)}

\item{...}{Control parameters for OSA method}
}
\value{
\code{data.frame} with OSA \emph{standardized} residuals
in column \code{residual}. Depending on the method the output may
also include OSA expected observation in column \code{mean}.
}
\description{
Calculate one-step-ahead (OSA) residuals for a latent variable
model. (\emph{Modified from TMB version to allow easier parallel
computations})
}
\keyword{internal}
