% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{vcov.mpmm}
\alias{vcov.mpmm}
\title{Calculate Variance-Covariance Matrix for a Fitted mpmm model}
\usage{
\method{vcov}{mpmm}(object, full = FALSE, ...)
}
\arguments{
\item{object}{a \dQuote{mpmm} fit}

\item{full}{return a full variance-covariance matrix?}

\item{\dots}{ignored, for method compatibility}
}
\value{
By default (\code{full==FALSE}).  If \code{full==TRUE}, a single square variance-covariance matrix for \emph{all} top-level model parameters (conditional and variance-covariance parameters)
}
\description{
Calculate Variance-Covariance Matrix for a Fitted mpmm model
}
