% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\name{anova.mpmm}
\alias{anova.mpmm}
\title{anova tables}
\usage{
\method{anova}{mpmm}(object, ...)
}
\arguments{
\item{object}{an mpmm fit object}

\item{...}{additional mpmm fit objects}
}
\description{
perform likelihood ratio tests on 2 or more mpmm fit objects
}
