% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_control.R
\name{inner_control}
\alias{inner_control}
\title{Control Values for \code{mpmm}}
\usage{
inner_control(...)
}
\arguments{
\item{...}{control parameters for the inner optimizer}
}
\value{
Returns a list with components:
 \item{\code{control}}{list of control parameters for inner optimizer}
}
\description{
\code{inner_control} allows specification of control parameters for the
inner optimization used by \code{mpmm}
}
\details{
See \code{\link{MakeADFun}} and \code{\link{newton}}
for details and available options. Adapted from S. Wotherspoon
\url{https://github.com/SWotherspoon/RWalc/blob/master/R/RWalc.R}
}
\examples{
fit <- mpmm(~ ice + (ice | id),
data = ellie.ice.short,
inner.control = inner_control(tol = 1e-03))
}
\seealso{
\code{\link{MakeADFun}}, \code{\link{newton}}
}
