\name{mpmm-package}
\alias{mpmm-package}
\docType{package}
\title{
Marginal Product Mixture Model for Relational Event Data
}
\description{
This package implements the Marginal Product Mixture Model (MPMM) for
relational event data.  Inference for the model is implemented via a
collapsed Gibbs sampler written in C.  Utility functions for processing
sets of relational events and displaying posterior predictive distributions.
}
\details{
\tabular{ll}{
Package: \tab mpmm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-01-10\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Built: \tab R 2.12.0; i486-pc-linux-gnu; 2010-11-01 07:23:54 UTC; unix\cr
}

Index:
\preformatted{
mpmm                    Method for fitting the marginal product mixture
                        model.
}
}
\author{
Christopher DuBois

Maintainer: Christopher DuBois <duboisc@ics.uci.edu>
}

\keyword{ package }
