% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmm.r
\name{mpmm}
\alias{mpmm}
\title{Move Persistence Mixed-Effects Model}
\usage{
mpmm(
  formula = NA,
  data = NULL,
  method = "ML",
  optim = c("nlminb", "optim"),
  control = NULL,
  optMeth = "BFGS",
  verbose = 2,
  model = "mpmm"
)
}
\arguments{
\item{formula}{a right-hand-side regression formula (no response variable)}

\item{data}{a data frame of observations (see details)}

\item{method}{method for maximising the log-likelihood ("ML" or "REML")}

\item{optim}{numerical optimizer to be used (nlminb or optim)}

\item{control}{a list of control parameters (currently only for nlminb)}

\item{optMeth}{outer optimisation method to be used: "BFGS" (unbounded) or "L-BFGS-B" (bounded)}

\item{verbose}{report progress during minimization (0 = silent (default); 1 = optimiser trace; 2 = parameter trace)}

\item{model}{"mpmm" or "mpmm_dt", "mpmm" is the default value and is for a model with regular time intervals between locations, "mpmm_dt" is for irregular time intervals.}
}
\value{
a list with components
\item{\code{states}}{a dataframe of estimated states}
\item{\code{fitted}}{a dataframe of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{input dataframe}
\item{\code{tmb}}{the tmb object}
\item{\code{opt}}{the object returned by the optimizer}
}
\description{
Fit a move persistence random walk via TMB to a pre-filtered/regularised animal
  track and estimate gamma as a linear function of covariates
}
\details{
The input track is given as a dataframe where each row is an
observed location and columns
\describe{
\item{'id'}{individual animal identifier,}
\item{'date'}{observation time (POSIXct,GMT),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude,}
\item{'tid'}{identifier for tracks if there are more than one track per individual (optional),}
\item{'...'}{named covariates appended to track}
}
}
\examples{
data(ellie.ice)
fit <- mpmm(~ ice + (1 | id), data = ellie.ice)
summary(fit)

}
