% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmm.r
\name{mpmm}
\alias{mpmm}
\title{Move Persistence Mixed-Effects Model}
\usage{
mpmm(
  formula = NA,
  data = NULL,
  map = NULL,
  se = TRUE,
  control = mpmm_control(),
  inner.control = inner_control(),
  model = "mpmm"
)
}
\arguments{
\item{formula}{a right-hand-side regression formula (no response variable)}

\item{data}{a data frame of observations (see details)}

\item{se}{whether to return standard errors}

\item{control}{a list of control parameters for the outer optimization
(see \code{\link{mpmm_control}})}

\item{inner.control}{a list of control parameters for the inner optimization
(see \code{\link{TMB::MakeADFun}} and \code{\link{TMB::newton}})}

\item{model}{"mpmm" or "mpmm_dt"; "mpmm" (default) for data with regular
time intervals between locations, "mpmm_dt" is for irregular time intervals}
}
\value{
a list with components
\item{\code{states}}{a dataframe of estimated states}
\item{\code{fitted}}{a dataframe of fitted locations}
\item{\code{par}}{model parameter summmary}
\item{\code{data}}{input dataframe}
\item{\code{tmb}}{the tmb object}
\item{\code{opt}}{the object returned by the optimizer}
}
\description{
Fit a move persistence random walk via TMB to a pre-filtered/regularised animal
  track and estimate gamma as a linear function of covariates
}
\details{
The input track is given as a dataframe where each row is an
observed location and columns
\describe{
\item{'id'}{individual animal identifier,}
\item{'date'}{observation time (POSIXct,GMT),}
\item{'lon'}{observed longitude,}
\item{'lat'}{observed latitude,}
\item{'tid'}{identifier for tracks if there are more than one track per
individual (optional),}
\item{'...'}{named covariates appended to track}
}
}
\examples{
data(ellie.ice)
fit <- mpmm(~ ice + (1 | id), data = ellie.ice)
summary(fit)

}
