% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mpmm.r
\name{plot.mpmm}
\alias{plot.mpmm}
\title{plot}
\usage{
\method{plot}{mpmm}(m, page = 1)
}
\arguments{
\item{page}{1 = plot all terms on a single page, 0 otherwise}

\item{x}{a fitted object of class mpmm}
}
\description{
Plot fixed and random components
}
