% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mpmm.r
\name{plot.mpmm}
\alias{plot.mpmm}
\title{plot}
\usage{
\method{plot}{mpmm}(x, label = FALSE, ...)
}
\arguments{
\item{x}{an mpmm fit object}

\item{label}{add id labels to random effects}

\item{...}{additional arguments to be ignored}
}
\description{
Visualise fixed and random covariate relationships from
an mpmm fit object
}
