% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mpmm.r
\name{plot.mpmm}
\alias{plot.mpmm}
\title{plot}
\usage{
\method{plot}{mpmm}(m, label = FALSE, page = 1)
}
\arguments{
\item{m}{a fitted object of class mpmm}

\item{page}{1 = plot all terms on a single page, 0 otherwise}
}
\description{
Visualise fixed and random relationships
}
